Pickup script aangepast
