+++ Assets/Scenes/Test-Tessa-Scene.unity
@@ -572,7 +572,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 182112406}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.45, y: 1.19, z: 8.03}
+  m_LocalPosition: {x: 0.22, y: 1.19, z: 8.03}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 598904879}
@@ -927,7 +927,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!114 &498318225
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1121,7 +1121,7 @@ Rigidbody:
   m_Drag: 0
   m_AngularDrag: 0.05
   m_UseGravity: 1
-  m_IsKinematic: 0
+  m_IsKinematic: 1
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
@@ -1242,9 +1242,9 @@ GameObject:
   - component: {fileID: 567473050}
   - component: {fileID: 567473049}
   - component: {fileID: 567473053}
+  - component: {fileID: 567473054}
   - component: {fileID: 567473048}
   - component: {fileID: 567473047}
-  - component: {fileID: 567473054}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -1279,6 +1279,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 10
+  playerSens: 500
 --- !u!136 &567473049
 CapsuleCollider:
   m_ObjectHideFlags: 0
@@ -1369,7 +1370,7 @@ Rigidbody:
   m_Drag: 0
   m_AngularDrag: 0.05
   m_UseGravity: 1
-  m_IsKinematic: 1
+  m_IsKinematic: 0
   m_Interpolate: 0
   m_Constraints: 112
   m_CollisionDetection: 0
@@ -1386,7 +1387,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   equipPosition: {fileID: 182112407}
-  distance: 10
+  distance: 5
 --- !u!1 &598904876
 GameObject:
   m_ObjectHideFlags: 0
