+++ Assets/Scripts/PickUpThings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8e4f28f541798614d9d59d45be2df6f5
+guid: eeb41889d3abf2c48b2112f7d0cd16c9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
